[Release] 1.0버전 출시

+++ Cider/Cider/Presentation/Mypage/SubViews/LevelView.swift
@@ -131,7 +131,7 @@ final class LevelView: UIView {
 extension LevelView {
 
     func setUp(
-        percent: Float,
+        percent: Double,
         experience: String,
         level: String,
         currentLevel: String,
